+++ packages/core/src/config/ConfigUtils.js
@@ -23,13 +23,15 @@ class ConfigUtils {
     config.dataPath = path.join(container, 'data')
     config.workingDir = '/'
     config.localWorkingDir = localWorkingDir
-    config.keysPath = path.join(container, 'keys.json')
+    config.keysDataPath = path.join(container, 'keys')
+    config.keysPath = path.join(config.keysDataPath, 'default.json')
 
     config.localDataPath = path.join(container, 'local')
     config.tmpPath = path.join(container, 'tmp')
 
     await fs.emptyDir(config.tmpPath)
     await fs.ensureDir(config.dataPath)
+    await fs.ensureDir(config.keysDataPath)
     await fs.ensureDir(config.localDataPath)
 
     let oldEnvPath = path.join(container, 'env.json')
