+++ packages/core/src/index.js
@@ -1,6 +1,5 @@
 module.exports = {
   Secrez: require('./Secrez'),
-  Crypto: require('@secrez/crypto'),
   Entry: require('./Entry'),
   config: require('./config'),
   ConfigUtils: require('./config/ConfigUtils'),
