+++ packages/courier/package.json
@@ -17,6 +17,7 @@
   },
   "dependencies": {
     "@secrez/core": "workspace:~0.9.0",
+    "@secrez/crypto": "workspace:~0.2.0",
     "@secrez/fs": "workspace:~0.9.0",
     "@secrez/hub": "workspace:~0.2.0",
     "@secrez/tls": "workspace:~0.1.4",
