+++ packages/crypto/test/index.test.js
@@ -347,6 +347,20 @@ describe('#Crypto', function () {
 
   })
 
+  describe('#fromTsToDate', async function () {
+
+    it('should recover a date from a timestamp with microseconds', async function () {
+
+      for (let i = 0; i < 20; i++) {
+        let ts = Crypto.getTimestampWithMicroseconds().join('.')
+        let d = (new Date).toISOString().substring(0, 18)
+        assert.isTrue(RegExp('^' + d).test(Crypto.fromTsToDate(ts)[0]))
+        sleep(1)
+      }
+    })
+
+  })
+
   describe.skip('performance comparision between V1 and V2', async function () {
 
     it('should compare encryption V1 and V2', async function() {
