+++ packages/fs/package.json
@@ -14,6 +14,7 @@
   },
   "dependencies": {
     "@secrez/core": "workspace:~0.9.0",
+    "@secrez/crypto": "workspace:~0.2.0",
     "@secrez/utils": "workspace:~0.1.6",
     "command-line-args": "^5.1.1",
     "fs-extra": "^8.1.0",
