+++ packages/fs/test/Node.test.js
@@ -2,7 +2,8 @@ const assert = require('chai').assert
 const fs = require('fs-extra')
 const util = require('util')
 const path = require('path')
-const {config, Crypto, Entry} = require('@secrez/core')
+const {config, Entry} = require('@secrez/core')
+const Crypto = require('@secrez/crypto')
 const Secrez = require('@secrez/core').Secrez(Math.random())
 const Node = require('../src/Node')
 const {jsonEqual, initRandomNode, setNewNodeVersion, initARootNode} = require('@secrez/test-helpers')
