+++ packages/hub/package.json
@@ -26,6 +26,7 @@
   "preferGlobal": true,
   "dependencies": {
     "@secrez/core": "workspace:~0.9.0",
+    "@secrez/crypto": "workspace:~0.2.0",
     "@secrez/tls": "workspace:~0.1.4",
     "@secrez/utils": "workspace:~0.1.6",
     "book": "1.3.3",
