+++ packages/migrate/package.json
@@ -16,10 +16,12 @@
   },
   "dependencies": {
     "@secrez/core": "workspace:~0.9.0",
+    "@secrez/crypto": "workspace:~0.2.0",
     "secrez-core-0-8-2": "npm:@secrez/core@0.8.12",
     "@secrez/fs": "workspace:~0.9.0",
     "secrez-fs-0-8-8": "npm:@secrez/fs@0.8.8",
     "@secrez/hub": "workspace:~0.2.0",
+    "secrez": "workspace:~0.11.0",
     "@secrez/utils": "workspace:~0.1.6",
     "case": "^1.6.3",
     "chalk": "^3.0.0",
