+++ packages/secrez/package.json
@@ -16,6 +16,7 @@
   },
   "dependencies": {
     "@secrez/core": "workspace:~0.9.0",
+    "@secrez/crypto": "workspace:~0.2.0",
     "@secrez/fs": "workspace:~0.9.0",
     "@secrez/hub": "workspace:~0.2.0",
     "@secrez/utils": "workspace:~0.1.6",
