+++ packages/secrez/src/PreCommand.js
@@ -1,5 +1,5 @@
 const {chalk} = require('./utils/Logger')
-const {Crypto} = require('@secrez/core')
+const Crypto = require('@secrez/crypto')
 const {utils: hubUtils} = require('@secrez/hub')
 const superagent = require('superagent')
