+++ packages/secrez/src/commands/Ls.js
@@ -166,6 +166,7 @@ class Ls extends require('../Command') {
         )
       }
     } catch (e) {
+      console.log(e)
       this.Logger.red(e.message)
     }
     await this.prompt.run()
