+++ packages/secrez/src/commands/Totp.js
@@ -200,7 +200,7 @@ class Totp extends require('../Command') {
               `Try it, running "totp ${node.getPath()}"`
             ].join('\n')
           } else {
-            let totp = parsed.totp.replace(/\s/g, '')
+            let totp = parsed.totp ? parsed.totp.replace(/\s/g, '') : null
             if (totp) {
               const token = authenticator.generate(totp)
               this.prompt.commands.copy.copy({
