+++ packages/secrez/src/index.js
@@ -1,3 +1,4 @@
 module.exports = {
-  version: require('../package').version
+  version: require('../package').version,
+  Prompt: require('./prompts/MainPromptMock')
 }
