+++ packages/tunnel/package.json
@@ -31,6 +31,7 @@
   },
   "dependencies": {
     "@secrez/core": "workspace:~0.9.0",
+    "@secrez/crypto": "workspace:~0.2.0",
     "@secrez/utils": "workspace:~0.1.6",
     "chalk": "^3.0.0",
     "command-line-args": "^5.1.1",
